server/requirements.txt
annotated-types==0.7.0
anyio==3.6.2
asgiref==3.8.1
attrs==20.3.0
autopep8==2.1.0
aws-lambda-powertools==1.21.0
aws-xray-sdk==2.14.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.38.2
botocore==1.38.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.3.2
check50==3.3.11
clang-format==18.1.4
click==8.1.3
cryptography==39.0.2
distlib==0.3.6
docker==7.1.0
ecdsa==0.18.0
EditorConfig==0.12.4
fastapi==0.68.0
fastjsonschema==2.21.1
filelock==3.9.0
Flask==3.1.0
h11==0.14.0
httptools==0.5.0
icdiff==2.0.7
idna==3.4
itsdangerous==2.2.0
jellyfish==0.11.2
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.4.2
jsbeautifier==1.15.1
lib50==3.0.11
libmagic==1.0
mangum==0.19.0
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
passlib==1.7.4
pexpect==4.9.0
pillow==10.3.0
pipenv==2023.2.18
pippip==0.0.10
platformdirs==3.1.0
ptyprocess==0.7.0
pyasn1==0.4.8
pycodestyle==2.11.1
pycparser==2.21
pydantic==1.10.22
pydantic_core==2.33.1
pygame==2.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.12.0
scipy==1.13.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.14.2
style50==2.10.3
submit50==3.1.4
termcolor==1.1.0
threadpoolctl==3.5.0
typing-inspection==0.4.0
typing_extensions==4.5.0
urllib3==2.2.1
uvicorn==0.15.0
uvloop==0.17.0
virtualenv==20.20.0
virtualenv-clone==0.5.7
watchfiles==0.18.1
websockets==10.4
Werkzeug==3.1.3
wrapt==1.17.2
